Only reload active tabs when enabling/disabling the mask
If you have many tabs open with the same domain, it's really a bad idea
to try reloading all of them at the same time. This still reloads it in
multiple windows (if you have multiple windows where the same website is
currently loaded in the active tab), but I think this is fine for now.